@@ -9,7 +9,7 @@ enum ThemePreference { light, dark, system }
 
 class ThemeProvider with ChangeNotifier {
   ThemePreference _themePreference = ThemePreference.system;
-  bool _isDarkMode = false; // This will now be a derived state
+  bool _isDarkMode = false;
   bool _useSystemFont = false;
 
   static const String _themePreferenceKey = 'themePreference';
@@ -96,73 +96,231 @@ class ThemeProvider with ChangeNotifier {
     return TextStyle(
       fontSize: fontSize,
       fontWeight: fontWeight,
-      color:
-          color ?? (_isDarkMode ? Colors.white.withAlpha(222) : Colors.black87),
+      color: color ?? (_isDarkMode ? Colors.white : Colors.black),
     );
   }
 
-  // --- PURE DARK THEME ---
+  // --- LIQUID GLASS DESIGN HELPERS ---
+
+  /// Returns the glass background color based on theme
+  Color get glassBackground => _isDarkMode
+      ? Colors.white.withOpacity(0.08)
+      : Colors.black.withOpacity(0.04);
+
+  /// Returns the glass border color based on theme
+  Color get glassBorder => _isDarkMode
+      ? Colors.white.withOpacity(0.15)
+      : Colors.black.withOpacity(0.08);
+
+  /// Returns a subtle glass highlight for top edges
+  Color get glassHighlight => _isDarkMode
+      ? Colors.white.withOpacity(0.2)
+      : Colors.white.withOpacity(0.7);
+
+  /// Returns card surface color for glass effect
+  Color get glassSurface =>
+      _isDarkMode ? const Color(0xFF0A0A0A) : const Color(0xFFF8F8F8);
+
+  /// Returns the elevated glass surface color
+  Color get glassElevatedSurface =>
+      _isDarkMode ? const Color(0xFF121212) : Colors.white;
+
+  // --- PURE BLACK DARK THEME ---
   ThemeData get darkTheme => ThemeData(
     brightness: Brightness.dark,
     scaffoldBackgroundColor: Colors.black,
-    cardColor: Colors.black,
+    canvasColor: Colors.black,
+    cardColor: const Color(0xFF0A0A0A),
     fontFamily: _useSystemFont ? null : 'Bebas',
+    primaryColor: Colors.white,
     appBarTheme: const AppBarTheme(
       backgroundColor: Colors.black,
       foregroundColor: Colors.white,
       elevation: 0,
       centerTitle: true,
+      surfaceTintColor: Colors.transparent,
       titleTextStyle: TextStyle(
         fontFamily: 'Bebas',
         fontSize: 22,
         fontWeight: FontWeight.bold,
+        color: Colors.white,
       ),
     ),
-    drawerTheme: const DrawerThemeData(backgroundColor: Color(0xFF1A1A1A)),
-    floatingActionButtonTheme: const FloatingActionButtonThemeData(
+    drawerTheme: const DrawerThemeData(
+      backgroundColor: Colors.black,
+      surfaceTintColor: Colors.transparent,
+    ),
+    floatingActionButtonTheme: FloatingActionButtonThemeData(
       backgroundColor: Colors.white,
       foregroundColor: Colors.black,
+      elevation: 0,
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
     ),
-    listTileTheme: const ListTileThemeData(iconColor: Colors.white70),
-    colorScheme: const ColorScheme.dark(
+    listTileTheme: const ListTileThemeData(
+      iconColor: Colors.white70,
+      textColor: Colors.white,
+    ),
+    iconTheme: const IconThemeData(color: Colors.white),
+    dividerColor: Colors.white.withOpacity(0.1),
+    colorScheme: ColorScheme.dark(
       primary: Colors.white,
       secondary: Colors.white,
-      surface: Color(0xFF1A1A1A),
+      surface: Colors.black,
+      surfaceContainerHighest: const Color(0xFF0A0A0A),
+      surfaceContainer: const Color(0xFF0A0A0A),
       onPrimary: Colors.black,
       onSecondary: Colors.black,
       onSurface: Colors.white,
-      error: Colors.redAccent,
+      error: Colors.red.shade400,
+      outline: Colors.white.withOpacity(0.15),
+      primaryContainer: Colors.white.withOpacity(0.1),
+      onPrimaryContainer: Colors.white,
     ),
     pageTransitionsTheme: _pageTransitionsTheme,
     inputDecorationTheme: InputDecorationTheme(
       filled: true,
-      fillColor: const Color(0xFF1A1A1A),
+      fillColor: Colors.white.withOpacity(0.05),
+      hintStyle: TextStyle(color: Colors.white.withOpacity(0.4)),
+      labelStyle: TextStyle(color: Colors.white.withOpacity(0.7)),
       border: OutlineInputBorder(
-        borderRadius: BorderRadius.circular(12),
-        borderSide: BorderSide(color: Colors.grey.shade800),
+        borderRadius: BorderRadius.circular(16),
+        borderSide: BorderSide(color: Colors.white.withOpacity(0.1)),
       ),
       enabledBorder: OutlineInputBorder(
-        borderRadius: BorderRadius.circular(12),
-        borderSide: BorderSide(color: Colors.grey.shade800),
+        borderRadius: BorderRadius.circular(16),
+        borderSide: BorderSide(color: Colors.white.withOpacity(0.1)),
+      ),
+      focusedBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(16),
+        borderSide: const BorderSide(color: Colors.white, width: 1.5),
       ),
     ),
     dialogTheme: DialogThemeData(
-      backgroundColor: const Color(0xFF1A1A1A),
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+      backgroundColor: const Color(0xFF0A0A0A),
+      surfaceTintColor: Colors.transparent,
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(24)),
+    ),
+    snackBarTheme: SnackBarThemeData(
+      backgroundColor: Colors.white,
+      contentTextStyle: const TextStyle(color: Colors.black),
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
+      behavior: SnackBarBehavior.floating,
+    ),
+    navigationBarTheme: NavigationBarThemeData(
+      backgroundColor: Colors.black,
+      surfaceTintColor: Colors.transparent,
+      indicatorColor: Colors.white.withOpacity(0.15),
+      iconTheme: WidgetStateProperty.resolveWith((states) {
+        if (states.contains(WidgetState.selected)) {
+          return const IconThemeData(color: Colors.white);
+        }
+        return IconThemeData(color: Colors.white.withOpacity(0.5));
+      }),
+      labelTextStyle: WidgetStateProperty.resolveWith((states) {
+        if (states.contains(WidgetState.selected)) {
+          return const TextStyle(
+            color: Colors.white,
+            fontSize: 12,
+            fontWeight: FontWeight.w600,
+          );
+        }
+        return TextStyle(color: Colors.white.withOpacity(0.5), fontSize: 12);
+      }),
+    ),
+    tabBarTheme: TabBarThemeData(
+      labelColor: Colors.white,
+      unselectedLabelColor: Colors.white.withOpacity(0.5),
+      indicatorColor: Colors.white,
+      dividerColor: Colors.transparent,
+    ),
+    switchTheme: SwitchThemeData(
+      thumbColor: WidgetStateProperty.resolveWith((states) {
+        if (states.contains(WidgetState.selected)) return Colors.black;
+        return Colors.white.withOpacity(0.6);
+      }),
+      trackColor: WidgetStateProperty.resolveWith((states) {
+        if (states.contains(WidgetState.selected)) return Colors.white;
+        return Colors.white.withOpacity(0.2);
+      }),
+    ),
+    elevatedButtonTheme: ElevatedButtonThemeData(
+      style: ElevatedButton.styleFrom(
+        backgroundColor: Colors.white,
+        foregroundColor: Colors.black,
+        elevation: 0,
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+        textStyle: const TextStyle(fontWeight: FontWeight.bold),
+      ),
+    ),
+    outlinedButtonTheme: OutlinedButtonThemeData(
+      style: OutlinedButton.styleFrom(
+        foregroundColor: Colors.white,
+        side: BorderSide(color: Colors.white.withOpacity(0.3)),
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+      ),
+    ),
+    textButtonTheme: TextButtonThemeData(
+      style: TextButton.styleFrom(foregroundColor: Colors.white),
+    ),
+    filledButtonTheme: FilledButtonThemeData(
+      style: FilledButton.styleFrom(
+        backgroundColor: Colors.white,
+        foregroundColor: Colors.black,
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+      ),
+    ),
+    segmentedButtonTheme: SegmentedButtonThemeData(
+      style: ButtonStyle(
+        backgroundColor: WidgetStateProperty.resolveWith((states) {
+          if (states.contains(WidgetState.selected)) return Colors.white;
+          return Colors.transparent;
+        }),
+        foregroundColor: WidgetStateProperty.resolveWith((states) {
+          if (states.contains(WidgetState.selected)) return Colors.black;
+          return Colors.white.withOpacity(0.6);
+        }),
+        side: WidgetStateProperty.all(
+          BorderSide(color: Colors.white.withOpacity(0.15)),
+        ),
+      ),
+    ),
+    chipTheme: ChipThemeData(
+      backgroundColor: Colors.white.withOpacity(0.08),
+      labelStyle: const TextStyle(color: Colors.white),
+      side: BorderSide(color: Colors.white.withOpacity(0.1)),
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
+    ),
+    popupMenuTheme: PopupMenuThemeData(
+      color: const Color(0xFF0A0A0A),
+      surfaceTintColor: Colors.transparent,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(16),
+        side: BorderSide(color: Colors.white.withOpacity(0.1)),
+      ),
+    ),
+    bottomSheetTheme: const BottomSheetThemeData(
+      backgroundColor: Colors.black,
+      surfaceTintColor: Colors.transparent,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.vertical(top: Radius.circular(24)),
+      ),
     ),
   );
 
-  // --- PURE LIGHT THEME (FIXED: Removed red accent) ---
+  // --- PURE WHITE LIGHT THEME ---
   ThemeData get lightTheme => ThemeData(
     brightness: Brightness.light,
     scaffoldBackgroundColor: Colors.white,
-    cardColor: Colors.white,
+    canvasColor: Colors.white,
+    cardColor: const Color(0xFFF5F5F5),
     fontFamily: _useSystemFont ? null : 'Bebas',
+    primaryColor: Colors.black,
     appBarTheme: const AppBarTheme(
       backgroundColor: Colors.white,
       foregroundColor: Colors.black,
       elevation: 0,
       centerTitle: true,
+      surfaceTintColor: Colors.transparent,
       titleTextStyle: TextStyle(
         fontFamily: 'Bebas',
         fontSize: 22,
@@ -170,37 +328,164 @@ class ThemeProvider with ChangeNotifier {
         color: Colors.black,
       ),
     ),
-    drawerTheme: const DrawerThemeData(backgroundColor: Colors.white),
-    floatingActionButtonTheme: const FloatingActionButtonThemeData(
-      backgroundColor: Colors.black, // Changed from red
-      foregroundColor: Colors.white,
+    drawerTheme: const DrawerThemeData(
+      backgroundColor: Colors.white,
+      surfaceTintColor: Colors.transparent,
     ),
-    listTileTheme: const ListTileThemeData(iconColor: Colors.black54),
+    floatingActionButtonTheme: FloatingActionButtonThemeData(
+      backgroundColor: Colors.black,
+      foregroundColor: Colors.white,
+      elevation: 0,
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+    ),
+    listTileTheme: const ListTileThemeData(
+      iconColor: Colors.black54,
+      textColor: Colors.black,
+    ),
+    iconTheme: const IconThemeData(color: Colors.black),
+    dividerColor: Colors.black.withOpacity(0.08),
     colorScheme: ColorScheme.light(
-      primary: Colors.black, // Changed from red
-      secondary: Colors.black, // Changed from red
+      primary: Colors.black,
+      secondary: Colors.black,
       surface: Colors.white,
+      surfaceContainerHighest: const Color(0xFFF5F5F5),
+      surfaceContainer: const Color(0xFFF8F8F8),
       onPrimary: Colors.white,
       onSecondary: Colors.white,
       onSurface: Colors.black,
-      error: Colors.red.shade900,
+      error: Colors.red.shade700,
+      outline: Colors.black.withOpacity(0.1),
+      primaryContainer: Colors.black.withOpacity(0.08),
+      onPrimaryContainer: Colors.black,
     ),
     pageTransitionsTheme: _pageTransitionsTheme,
     inputDecorationTheme: InputDecorationTheme(
       filled: true,
-      fillColor: Colors.white,
+      fillColor: Colors.black.withOpacity(0.03),
+      hintStyle: TextStyle(color: Colors.black.withOpacity(0.4)),
+      labelStyle: TextStyle(color: Colors.black.withOpacity(0.6)),
       border: OutlineInputBorder(
-        borderRadius: BorderRadius.circular(12),
-        borderSide: BorderSide(color: Colors.grey.shade300),
+        borderRadius: BorderRadius.circular(16),
+        borderSide: BorderSide(color: Colors.black.withOpacity(0.08)),
       ),
       enabledBorder: OutlineInputBorder(
-        borderRadius: BorderRadius.circular(12),
-        borderSide: BorderSide(color: Colors.grey.shade300),
+        borderRadius: BorderRadius.circular(16),
+        borderSide: BorderSide(color: Colors.black.withOpacity(0.08)),
+      ),
+      focusedBorder: OutlineInputBorder(
+        borderRadius: BorderRadius.circular(16),
+        borderSide: const BorderSide(color: Colors.black, width: 1.5),
       ),
     ),
     dialogTheme: DialogThemeData(
       backgroundColor: Colors.white,
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+      surfaceTintColor: Colors.transparent,
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(24)),
+    ),
+    snackBarTheme: SnackBarThemeData(
+      backgroundColor: Colors.black,
+      contentTextStyle: const TextStyle(color: Colors.white),
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
+      behavior: SnackBarBehavior.floating,
+    ),
+    navigationBarTheme: NavigationBarThemeData(
+      backgroundColor: Colors.white,
+      surfaceTintColor: Colors.transparent,
+      indicatorColor: Colors.black.withOpacity(0.1),
+      iconTheme: WidgetStateProperty.resolveWith((states) {
+        if (states.contains(WidgetState.selected)) {
+          return const IconThemeData(color: Colors.black);
+        }
+        return IconThemeData(color: Colors.black.withOpacity(0.4));
+      }),
+      labelTextStyle: WidgetStateProperty.resolveWith((states) {
+        if (states.contains(WidgetState.selected)) {
+          return const TextStyle(
+            color: Colors.black,
+            fontSize: 12,
+            fontWeight: FontWeight.w600,
+          );
+        }
+        return TextStyle(color: Colors.black.withOpacity(0.4), fontSize: 12);
+      }),
+    ),
+    tabBarTheme: TabBarThemeData(
+      labelColor: Colors.black,
+      unselectedLabelColor: Colors.black.withOpacity(0.4),
+      indicatorColor: Colors.black,
+      dividerColor: Colors.transparent,
+    ),
+    switchTheme: SwitchThemeData(
+      thumbColor: WidgetStateProperty.resolveWith((states) {
+        if (states.contains(WidgetState.selected)) return Colors.white;
+        return Colors.black.withOpacity(0.4);
+      }),
+      trackColor: WidgetStateProperty.resolveWith((states) {
+        if (states.contains(WidgetState.selected)) return Colors.black;
+        return Colors.black.withOpacity(0.15);
+      }),
+    ),
+    elevatedButtonTheme: ElevatedButtonThemeData(
+      style: ElevatedButton.styleFrom(
+        backgroundColor: Colors.black,
+        foregroundColor: Colors.white,
+        elevation: 0,
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+        textStyle: const TextStyle(fontWeight: FontWeight.bold),
+      ),
+    ),
+    outlinedButtonTheme: OutlinedButtonThemeData(
+      style: OutlinedButton.styleFrom(
+        foregroundColor: Colors.black,
+        side: BorderSide(color: Colors.black.withOpacity(0.2)),
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+      ),
+    ),
+    textButtonTheme: TextButtonThemeData(
+      style: TextButton.styleFrom(foregroundColor: Colors.black),
+    ),
+    filledButtonTheme: FilledButtonThemeData(
+      style: FilledButton.styleFrom(
+        backgroundColor: Colors.black,
+        foregroundColor: Colors.white,
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16)),
+      ),
+    ),
+    segmentedButtonTheme: SegmentedButtonThemeData(
+      style: ButtonStyle(
+        backgroundColor: WidgetStateProperty.resolveWith((states) {
+          if (states.contains(WidgetState.selected)) return Colors.black;
+          return Colors.transparent;
+        }),
+        foregroundColor: WidgetStateProperty.resolveWith((states) {
+          if (states.contains(WidgetState.selected)) return Colors.white;
+          return Colors.black.withOpacity(0.5);
+        }),
+        side: WidgetStateProperty.all(
+          BorderSide(color: Colors.black.withOpacity(0.1)),
+        ),
+      ),
+    ),
+    chipTheme: ChipThemeData(
+      backgroundColor: Colors.black.withOpacity(0.05),
+      labelStyle: const TextStyle(color: Colors.black),
+      side: BorderSide(color: Colors.black.withOpacity(0.08)),
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
+    ),
+    popupMenuTheme: PopupMenuThemeData(
+      color: Colors.white,
+      surfaceTintColor: Colors.transparent,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(16),
+        side: BorderSide(color: Colors.black.withOpacity(0.08)),
+      ),
+    ),
+    bottomSheetTheme: const BottomSheetThemeData(
+      backgroundColor: Colors.white,
+      surfaceTintColor: Colors.transparent,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.vertical(top: Radius.circular(24)),
+      ),
     ),
   );
 }
