@@ -1,5 +1,8 @@
+import 'dart:ui';
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
 import 'package:wallet/models/dataentry.dart';
+import 'package:wallet/models/theme_provider.dart';
 
 class AddCardScreen extends StatefulWidget {
   const AddCardScreen({super.key});
@@ -26,25 +29,98 @@ class _AddCardScreenState extends State<AddCardScreen>
 
   @override
   Widget build(BuildContext context) {
+    final themeProvider = Provider.of<ThemeProvider>(context);
+    final isDark = themeProvider.isDarkMode;
+    final textColor = isDark ? Colors.white : Colors.black;
+
     return Scaffold(
       appBar: AppBar(
         title: const Text('Add a New Card'),
-        bottom: TabBar(
-          controller: _tabController,
-          tabs: const [
-            Tab(icon: Icon(Icons.credit_card), text: 'Credit/Debit'),
-            Tab(icon: Icon(Icons.shopping_basket_outlined), text: 'Loyalty'),
-            Tab(icon: Icon(Icons.fingerprint), text: 'Identity'),
-          ],
+        leading: Container(
+          margin: const EdgeInsets.all(8),
+          decoration: BoxDecoration(
+            color: isDark
+                ? Colors.white.withOpacity(0.08)
+                : Colors.black.withOpacity(0.05),
+            borderRadius: BorderRadius.circular(12),
+          ),
+          child: IconButton(
+            icon: Icon(
+              Icons.arrow_back_ios_new_rounded,
+              color: textColor,
+              size: 20,
+            ),
+            onPressed: () => Navigator.pop(context),
+          ),
+        ),
+        bottom: PreferredSize(
+          preferredSize: const Size.fromHeight(60),
+          child: Container(
+            margin: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+            decoration: BoxDecoration(
+              borderRadius: BorderRadius.circular(16),
+              color: isDark
+                  ? Colors.white.withOpacity(0.06)
+                  : Colors.black.withOpacity(0.03),
+              border: Border.all(
+                color: isDark
+                    ? Colors.white.withOpacity(0.1)
+                    : Colors.black.withOpacity(0.05),
+              ),
+            ),
+            child: ClipRRect(
+              borderRadius: BorderRadius.circular(16),
+              child: BackdropFilter(
+                filter: ImageFilter.blur(sigmaX: 10, sigmaY: 10),
+                child: TabBar(
+                  controller: _tabController,
+                  indicator: BoxDecoration(
+                    borderRadius: BorderRadius.circular(12),
+                    color: isDark ? Colors.white : Colors.black,
+                  ),
+                  indicatorSize: TabBarIndicatorSize.tab,
+                  indicatorPadding: const EdgeInsets.all(4),
+                  labelColor: isDark ? Colors.black : Colors.white,
+                  unselectedLabelColor: textColor.withOpacity(0.5),
+                  labelStyle: const TextStyle(
+                    fontWeight: FontWeight.w600,
+                    fontSize: 12,
+                  ),
+                  unselectedLabelStyle: const TextStyle(
+                    fontWeight: FontWeight.normal,
+                    fontSize: 12,
+                  ),
+                  dividerColor: Colors.transparent,
+                  tabs: [
+                    _buildTab(Icons.credit_card_rounded, 'Credit/Debit'),
+                    _buildTab(Icons.shopping_basket_rounded, 'Loyalty'),
+                    _buildTab(Icons.fingerprint_rounded, 'Identity'),
+                  ],
+                ),
+              ),
+            ),
+          ),
         ),
       ),
       body: TabBarView(
         controller: _tabController,
         children: const [
-          // These are the refactored form widgets from dataentry.dart
           CreditCardEntryForm(),
           BarcodeCardEntryForm(cardType: BarcodeCardType.loyalty),
           BarcodeCardEntryForm(cardType: BarcodeCardType.identity),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildTab(IconData icon, String text) {
+    return Tab(
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Icon(icon, size: 18),
+          const SizedBox(width: 6),
+          Flexible(child: Text(text, overflow: TextOverflow.ellipsis)),
         ],
       ),
     );
